backend/requirements.txt
langchain
sentence-transformers
fastapi
uvicorn
pypdf
python-dotenv
pinecone[grpc]
langchain-pinecone
langchain-community
langchain-groq
langchain-experimental
langchain-huggingface